% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3d_utils.R
\name{mclapply.hack}
\alias{mclapply.hack}
\title{mclapply.hack}
\usage{
mclapply.hack(..., verbose = FALSE, mc.cores = 1)
}
\arguments{
\item{...}{arguments as for lapply}

\item{verbose}{bool}

\item{mc.cores}{integer number of cores}
}
\description{
This function mimics forking (mclapply) also for Windows.
}
\keyword{internal}
