% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3d_methods.R
\name{summary.r3d}
\alias{summary.r3d}
\title{Summary method for r3d objects}
\usage{
\method{summary}{r3d}(object, samples = c(0.25, 0.5, 0.75), ...)
}
\arguments{
\item{object}{an r3d object}

\item{samples}{numeric vector of quantile sub-ranges for aggregated summary. Defaults c(0.25, 0.5, 0.75).}

\item{...}{not used}
}
\description{
prints basic info, bandwidths, and possibly aggregated distribution effects
}
