% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3d_utils.R
\name{.compute_empirical_qmat}
\alias{.compute_empirical_qmat}
\title{Internal Utilities for R3D}
\usage{
.compute_empirical_qmat(Y_list, q_grid)
}
\description{
\itemize{
\item .compute_empirical_qmat: convert Y_list into a matrix of quantiles
\item ...
}
}
\keyword{internal}
