% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3d_methods.R
\name{plot.r3d}
\alias{plot.r3d}
\title{Plot method for r3d objects}
\usage{
\method{plot}{r3d}(x, main = NULL, xlab = "Quantile", ylab = "Treatment Effect", ...)
}
\arguments{
\item{x}{r3d object}

\item{main}{Title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{...}{not used}
}
\description{
Plots \tau(q) vs q with optional confidence bands
}
